Move logic for running user-defined code to a new class
